New OTDigitalDisplay component.
This can mostly be dropped in for a OTDataCollector and instead of showing a graph it just shows the current value.  It stores the collected data in a datastore, so that data can also be displayed in a table or a later graph.   It currently allows restarting of previously collected data.

git-svn-id: https://svn.concord.org/svn/projects/trunk/common/java/data/data@24909 6e01202a-0783-4428-890a-84243c50cc2b